+++ Projects/Feature/Onboarding/Interface/Sources/View/Onboarding/OnboardingCompletedView.swift
@@ -7,7 +7,6 @@
 
 import SwiftUI
 
-
 public struct OnboardingCompletedView: View {
     @Environment(StartUpViewModel.self) var startUpViewModel
     @Environment(OnboardingCompletedViewModel.self) var viewModel
